+++ Frontend/sky_line/app/Post/page.module.css
@@ -1,0 +1,65 @@
+/* AppPost.css */
+
+.create-post-button {
+    position: fixed;
+    bottom: 40px;
+    right: 40px;
+    width: 200px;
+    height: 60px;
+    background: linear-gradient(to right, blue, rgb(134, 148, 189));
+    border-radius: 30px;
+    border: none;
+    color: white;
+    display: flex;
+    justify-content: space-around;
+    align-items: center;
+    font-size: 1.2em;
+    font-weight: bold;
+    cursor: pointer;
+    padding: 10px;
+    box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.5);
+    overflow: hidden;
+    position: fixed;
+    transition: background-position 0.5s; /* Added transition for smoother effect */
+}
+
+.create-post-button::before {
+    content: '';
+    position: absolute;
+    top: 0;
+    left: -100%; /* Start from the left edge */
+    width: 100%;
+    height: 100%;
+    background: linear-gradient(to right, transparent, rgba(255, 255, 255, 0.1), transparent); /* Gradient for tide effect */
+    transition: left 0.5s; /* Added transition for smoother effect */
+}
+
+.create-post-button:hover::before {
+    left: 100%; /* Move the gradient to the right */
+}
+
+.create-post-button .create-icon {
+    margin-left: 10px; /* Adjust the margin as needed */
+}
+
+.modal-enter {
+    opacity: 0;
+    transform: translateY(100%);
+}
+
+.modal-enter-active {
+    opacity: 1;
+    transform: translateY(0);
+    transition: opacity 500ms, transform 500ms;
+}
+
+.modal-exit {
+    opacity: 1;
+    transform: translateY(0);
+}
+
+.modal-exit-active {
+    opacity: 0;
+    transform: translateY(100%);
+    transition: opacity 500ms, transform 500ms;
+}
