+++ Frontend/sky_line/app/Services/PublishPostService.tsx
@@ -1,0 +1,29 @@
+import axios, { Axios } from "axios";
+import Post from "../Post/page";
+
+const apiUrl = 'http://localhost:8080';
+
+const publishPostRequest = async (post : FormData) => {
+
+    const url = `${apiUrl}/api/publish_post`
+
+    const makeRequest = async(url: string, post : FormData) =>{
+        try {
+            const response = await axios.post(url, post);
+            return response.data;
+        } catch (error) {
+            console.error(error);
+            alert("Post is not published")
+            return "Post is not published";
+        }
+    }
+
+    const sendRequest = async () =>{
+        const response = await makeRequest(url, post);
+        return response; 
+    }
+
+  return sendRequest()
+};
+
+export default publishPostRequest;
